Normalized on use of modifiers instead of operators. Doc pass on side effects modifier for providers.
