+++ Sources/ResourceProvider/Cache/Modifiers/CacheMapValue.swift
