+++ Sources/ResourceProvider/Cache/Modifiers/CacheSerialized.swift
@@ -42,7 +42,7 @@ public extension SyncCache where Self: Sendable, ID: Sendable, Value: Sendable {
      Returns a wrapper for a ``SyncCache`` `& Sendable` that guarantees serialization.
 
      If a ``SyncCache`` needs to be used in an `async` context and it doesn't play well with reentrancy —usually
-     because you want to avoid data races with its state management— you will want to use this operator to make an
+     because you want to avoid data races with its state management— you will want to use this modifier to make an
      ``AsyncCache`` out of it.
 
      This is not particularly problematic for very fast caches i.e. in-memory ones. Normally you will be using a
