+++ Sources/ResourceProvider/Provider/Modifiers/Concurrent.swift
@@ -18,7 +18,7 @@ public extension SyncProvider where Self: Sendable {
      Returns a wrapper for a ``SyncProvider`` `& Sendable` that guarantees serialization.
 
      If a ``SyncProvider`` needs to be used in an concurrent context and it plays well with reentrancy you will want to
-     use this operator to make it into an ``AsyncProvider``.
+     use this modifier to make it into an ``AsyncProvider``.
 
      While you can sidestep the `Sendable` requirement by using ``forceSendable()`` on a non-`Sendable` ``SyncProvider``
      you should be very sure that it will behave properly in a concurrent context. If you can't guarantee reentrance
