+++ Sources/ResourceProvider/Provider/Modifiers/ForceSendable.swift
@@ -24,7 +24,7 @@ extension SyncProvider {
      force-converts a non-sendable synchronous provider into a sendable one, usually so you can apply ``concurrent()``
      or ``serialized()`` to it afterwards.
 
-     The operator can only be applied when both `ID` and `Value` are already `Sendable`. Use ``mapID(_:)-34d6t`` and
+     The modifier can only be applied when both `ID` and `Value` are already `Sendable`. Use ``mapID(_:)-34d6t`` and
      ``mapValue(_:)-3cj8y`` to achieve either if needed —if it involves `@unchecked Sendable` wrappers that's on you,
      the developer—.
      - Returns: An IKWID ``SyncProvider`` that has the exact same behavior as the caller but is seen by the compiler as
