+++ Sources/ResourceProvider/Provider/Modifiers/Serialized.swift
@@ -33,7 +33,7 @@ public extension SyncProvider where Self: Sendable, ID: Sendable, Value: Sendabl
      Returns a wrapper for a ``SyncProvider`` `& Sendable` that guarantees serialization.
 
      If a ``SyncProvider`` needs to be used in an `async` context and it doesn't play well with reentrancy —usually
-     because you want to avoid data races with its state management— you will want to use this operator to make an
+     because you want to avoid data races with its state management— you will want to use this modifier to make an
      ``AsyncCache`` out of it.
 
      This is not particularly problematic for very fast providers i.e. generative ones that don't take long and require
